requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-database-url==2.2.0
psycopg[binary]==3.2.1
gunicorn==22.0.0
whitenoise==6.7.0
django-cors-headers==4.4.0
python-dotenv==1.0.1